// FrontFredy/src/app/model/claim-detaill.ts
export class ClaimDetail {
	numberOfPieces: Number;
	weight: Number;
	weightUnit: String;
	volume: Number;
	volumeUnit: String;
	portOfDelivery: String;
	stationResposible: String;
	claimAmount: Number;
	claimAmountCurrency: String;
	approvedAmount: String;
	approvedAmountCurrency: String;
	reasonCode: String;
	reasonCodeId: String;
	commodityCode: String;
	claimDetailsSerialNumber: String;
	severity: String;
	investigationResults: String;
	approvalCateogry: String;
	operationalFlag: String;
	claimRemarks: String;
	limitDay:Number;
}