// FrontFredy/src/app/model/claim-summary.ts
import { AvDateTimeOffsetPipe } from "../pipes/av-date-time-offset.pipe";
import { ClaimAttachment } from "./claim-attachment.model";
import { ClaimDetail } from "./claim-detaill";

export class ClaimSummary {
  id: Number;
  claimReferenceNumber: String;
  claimantCode: String;
  stationCode: String;
  shipmentPrefix: String;
  documentOwnerId: String;
  masterDocumentNumber: String;
  sequenceNumber: String;
  duplicateNumber: String;
  companyCode: String;
  claimDate: String;
  nature: String;
  communicationMode: String;
  claimStatus: String;
  totalClaimAmount: Number;
  totalClaimAmountCurrency: String;
  totalApprovedAmount: Number;
  totalApprovedAmountCurrency: String;
  carrierLiabilityAmount: String;
  carrierLiabilityAmountCurrency: String;
  claimantName: String;
  claimantAddress1: String;
  daughterDocumentNumber: String;
  claimantAddress2: String;
  claimantZipCode: String;
  claimantCountry: String;
  claimantTelephoneNumber: String;
  claimantFax: string;
  claimantEmail: String;
  claimDetails: ClaimDetail[];
  claimAttachments: ClaimAttachment[];
  claimDescription: String;
  otherAmount: string;
  totalAmount: string;
  remarks: string;
  policieAccepted: boolean;
  policieAcceptedDate: Date;
  sourceSystem: String;
  createdDate: Date;
  //invoiceNumber: String;
  arrivalDate: Date;
  preeliminarDate: Date;
  formalDate: Date;
}