// FrontFredy/src/app/model/flight-booking.model.ts
export class FlightBooking {
  flightCarrierCode: string;
  flightNumber: Number;
  flightDate: Date;
  flightSegmentNumber: Number;
  flightSegmentOrigin: Number;
  flightSegmentDestination: Number;
  departureAiportCode: string;
  departureDate: Date;
  departureDateUtc: Date;
  departureDateType: string;
  arrivalAiportCode: string;
  arrivalDate: Date;
  arrivalDateUtc: Date;
  arrivalDateType: string;
}